services/geoai-service/requirements.txt
geoai-py
Flask
flask-cors
google-generativeai
python-dotenv
Pillow
